--- src/app/admin/change/page.tsx
@@ -1,6 +1,0 @@
-import { Change } from "@/components/AdminPage/Change";
-
-export default function ChangePage() {
-  return <Change />;
-}
-
